Validation result paso a middleware en register
